+++ SkillShare.Domain/Entities/StudentAnswer.cs
@@ -2,13 +2,16 @@
 
 namespace SkillShare.Domain.Entities;
 
+/// <summary>
+/// Сущность для работы с ответами студентов на вопросы
+/// </summary>
 public class StudentAnswer : IEntityId<long>, IAuditable
 {
     public long Id { get; set; }
 
     public long? TeacherId { get; set; }
 
-    public long QuestionId { get; set; }
+    public int QuestionId { get; set; }
 
     public long StudentId { get; set; }
